etag decorator that works with arbitrary objects created by views.
See https://docs.djangoproject.com/en/dev/topics/conditional-view-processing/ for details about why/how this might be a good idea.

This allows for shortcutting running the actual view function if we have a way of generating the Etag, and comparing.

This can also be used for conditional PUT/POST/DELETE requests.